+++ src/Journalist.WindowsAzure.Storage.Abstractions/Tables/BatchOperationExtensions.cs
@@ -11,18 +11,11 @@ namespace Journalist.WindowsAzure.Storage.Tables
             batchOperation.Insert(partitionKey, rowKey, EmptyDictionary.Get<string, object>());
         }
 
-        public static void InsertOrMerge(this IBatchOperation batchOperation, string partitionKey, string rowKey)
+        public static void Insert(this IBatchOperation batchOperation, string partitionKey)
         {
             Require.NotNull(batchOperation, "batchOperation");
 
-            batchOperation.InsertOrMerge(partitionKey, rowKey, EmptyDictionary.Get<string, object>());
-        }
-
-        public static void InsertOrReplace(this IBatchOperation batchOperation, string partitionKey, string rowKey)
-        {
-            Require.NotNull(batchOperation, "batchOperation");
-
-            batchOperation.InsertOrReplace(partitionKey, rowKey, EmptyDictionary.Get<string, object>());
+            batchOperation.Insert(partitionKey, EmptyDictionary.Get<string, object>());
         }
 
         public static void Merge(this IBatchOperation batchOperation, string partitionKey, string rowKey, string etag)
@@ -32,11 +25,53 @@ namespace Journalist.WindowsAzure.Storage.Tables
             batchOperation.Merge(partitionKey, rowKey, etag, EmptyDictionary.Get<string, object>());
         }
 
+        public static void Merge(this IBatchOperation batchOperation, string partitionKey, string etag)
+        {
+            Require.NotNull(batchOperation, "batchOperation");
+
+            batchOperation.Merge(partitionKey, etag, EmptyDictionary.Get<string, object>());
+        }
+
         public static void Replace(this IBatchOperation batchOperation, string partitionKey, string rowKey, string etag)
         {
             Require.NotNull(batchOperation, "batchOperation");
 
             batchOperation.Replace(partitionKey, rowKey, etag, EmptyDictionary.Get<string, object>());
+        }
+
+        public static void Replace(this IBatchOperation batchOperation, string partitionKey, string etag)
+        {
+            Require.NotNull(batchOperation, "batchOperation");
+
+            batchOperation.Replace(partitionKey, etag, EmptyDictionary.Get<string, object>());
+        }
+
+        public static void InsertOrMerge(this IBatchOperation batchOperation, string partitionKey, string rowKey)
+        {
+            Require.NotNull(batchOperation, "batchOperation");
+
+            batchOperation.InsertOrMerge(partitionKey, rowKey, EmptyDictionary.Get<string, object>());
+        }
+
+        public static void InsertOrMerge(this IBatchOperation batchOperation, string partitionKey)
+        {
+            Require.NotNull(batchOperation, "batchOperation");
+
+            batchOperation.InsertOrMerge(partitionKey, EmptyDictionary.Get<string, object>());
+        }
+
+        public static void InsertOrReplace(this IBatchOperation batchOperation, string partitionKey, string rowKey)
+        {
+            Require.NotNull(batchOperation, "batchOperation");
+
+            batchOperation.InsertOrReplace(partitionKey, rowKey, EmptyDictionary.Get<string, object>());
+        }
+
+        public static void InsertOrReplace(this IBatchOperation batchOperation, string partitionKey)
+        {
+            Require.NotNull(batchOperation, "batchOperation");
+
+            batchOperation.InsertOrReplace(partitionKey, EmptyDictionary.Get<string, object>());
         }
     }
 }