+++ src/Journalist.WindowsAzure.Storage.Abstractions/Tables/ICloudTableEntityRangeQuery.cs
@@ -5,6 +5,6 @@ namespace Journalist.WindowsAzure.Storage.Tables
 {
     public interface ICloudTableEntityRangeQuery
     {
-        Task<IList<IDictionary<string, object>>> ExecuteAsync();
+        Task<List<Dictionary<string, object>>> ExecuteAsync();
     }
 }