+++ src/Journalist.WindowsAzure.Storage/Tables/CloudTablePointQuery.cs
@@ -14,19 +14,18 @@ namespace Journalist.WindowsAzure.Storage.Tables
             string partitionKey,
             string rowKey,
             string[] properties,
-            Func<TableQuery<DynamicTableEntity>, TableContinuationToken, Task<TableQuerySegment<DynamicTableEntity>>>
-                fetchEntities,
+            Func<TableQuery<DynamicTableEntity>, TableContinuationToken, Task<TableQuerySegment<DynamicTableEntity>>> fetchEntities,
             ITableEntityConverter tableEntityConverter)
             : base(1, properties, fetchEntities, tableEntityConverter)
         {
             Require.NotEmpty(partitionKey, "partitionKey");
-            Require.NotEmpty(rowKey, "rowKey");
+            Require.NotNull(rowKey, "rowKey");
 
             m_partitionKey = partitionKey;
             m_rowKey = rowKey;
         }
 
-        public async Task<IDictionary<string, object>> ExecuteAsync()
+        public async Task<Dictionary<string, object>> ExecuteAsync()
         {
             var entities = await FetchEntities(
                 TableQuery.CombineFilters(